workflow fix?

+++ lecture/lecture3-4/functions_to_implement.cpp
@@ -35,7 +35,7 @@ std::vector<bool> OddMask(std::vector<int>);
 // Sums all numbers in a vector and returns the resulting value
 int Sum(std::vector<int> nums){
     int sum = 0;
-    for (int i = 0; i < nums.size(); i++){
+    for (unsigned long int i = 0; i < nums.size(); i++){
         sum += nums[i];
     }
